pyproject.toml
[tool.poetry]
name = "taskonomics"
version = "0.1.0"
description = "Taskonomics: Where Tasks Mean Business"
authors = ["Advik <adviksingh6632@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.15.0"
black = "^23.9.1"
ruff = "^0.0.291"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 120
